+++ oneview/resource_server_profile_template.go
@@ -16,6 +16,8 @@ import (
 	"github.com/HewlettPackard/oneview-golang/ov"
 	"github.com/HewlettPackard/oneview-golang/utils"
 	"github.com/hashicorp/terraform/helper/schema"
+	//	"encoding/json"
+	//	"io/ioutil"
 )
 
 func resourceServerProfileTemplate() *schema.Resource {
@@ -46,17 +48,13 @@ func resourceServerProfileTemplate() *schema.Resource {
 					Schema: map[string]*schema.Schema{
 						"manage_mode": {
 							Type:     schema.TypeBool,
-							Required: true,
+							Optional: true,
 						},
 						"mode": {
 							Type:     schema.TypeString,
-							Required: true,
+							Optional: true,
 						},
 						"pxe_boot_policy": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"secure_boot": {
 							Type:     schema.TypeString,
 							Optional: true,
 						},
@@ -308,24 +306,126 @@ func resourceServerProfileTemplate() *schema.Resource {
 							Optional: true,
 						},
 						"initialize": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-					},
-				},
-			},
-			"logical_drives": {
-				Optional: true,
-				Type:     schema.TypeSet,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"bootable": {
 							Type:     schema.TypeBool,
 							Optional: true,
 						},
-						"raid_level": {
-							Type:     schema.TypeString,
+						"controller": {
 							Optional: true,
+							Type:     schema.TypeSet,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"device_slot": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"initialize": {
+										Type:     schema.TypeBool,
+										Optional: true,
+									},
+									"drive_write_cache": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"mode": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"predictive_spare_rebuild": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"logical_drives": {
+										Optional: true,
+										Type:     schema.TypeSet,
+										Elem: &schema.Resource{
+											Schema: map[string]*schema.Schema{
+												"bootable": {
+													Type:     schema.TypeBool,
+													Optional: true,
+												},
+												"accelerator": {
+													Type:     schema.TypeString,
+													Optional: true,
+												},
+												"drive_technology": {
+													Type:     schema.TypeString,
+													Optional: true,
+												},
+												"name": {
+													Type:     schema.TypeString,
+													Optional: true,
+												},
+												"num_physical_drives": {
+													Type:     schema.TypeInt,
+													Optional: true,
+												},
+												"num_spare_drives": {
+													Type:     schema.TypeInt,
+													Optional: true,
+												},
+												"sas_logical_jbod_id": {
+													Type:     schema.TypeInt,
+													Optional: true,
+												},
+												"raid_level": {
+													Type:     schema.TypeString,
+													Optional: true,
+												},
+											},
+										},
+									},
+								},
+							},
+						},
+						"sas_logical_jbod": {
+							Optional: true,
+							Type:     schema.TypeSet,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"description": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"drive_slot": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"drive_max_size_gb": {
+										Type:     schema.TypeInt,
+										Optional: true,
+									},
+									"drive_min_size_gb": {
+										Type:     schema.TypeInt,
+										Optional: true,
+									},
+									"drive_technology": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"erase_data": {
+										Type:     schema.TypeBool,
+										Optional: true,
+									},
+									"id": {
+										Type:     schema.TypeInt,
+										Optional: true,
+									},
+									"name": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+									"num_physical_drive": {
+										Type:     schema.TypeInt,
+										Optional: true,
+									},
+									"persistent": {
+										Type:     schema.TypeBool,
+										Optional: true,
+									},
+								},
+							},
 						},
 					},
 				},
@@ -705,7 +805,6 @@ func resourceServerProfileTemplateCreate(d *schema.ResourceData, meta interface{
 			ManageMode:    rawBootMode["manage_mode"].(bool),
 			Mode:          rawBootMode["mode"].(string),
 			PXEBootPolicy: utils.Nstring(rawBootMode["pxe_boot_policy"].(string)),
-			SecureBoot:    rawBootMode["secure_boot"].(string),
 		}
 
 	}
@@ -761,28 +860,68 @@ func resourceServerProfileTemplateCreate(d *schema.ResourceData, meta interface{
 		}
 		serverProfileTemplate.Firmware = firmware
 	}
+
 	// Get local storage data if provided
 	rawLocalStorage := d.Get("local_storage").(*schema.Set).List()
 	localStorage := ov.LocalStorageOptions{}
 	for _, raw := range rawLocalStorage {
 		localStorageItem := raw.(map[string]interface{})
+		// Gets Local Storage Controller body
+		rawLocalStorageController := localStorageItem["controller"].(*schema.Set).List()
+		localStorageEmbeddedController := make([]ov.LocalStorageEmbeddedController, 0)
+		for _, raw2 := range rawLocalStorageController {
+			controller_data := raw2.(map[string]interface{})
+			// Gets Local Storage Controller's Logical Drives
+			rawLogicalDrives := controller_data["logical_drives"].(*schema.Set).List()
+			logicalDrives := make([]ov.LogicalDriveV3, 0)
+			for _, rawLogicalDrive := range rawLogicalDrives {
+				logicalDrivesItem := rawLogicalDrive.(map[string]interface{})
+				logicalDrives = append(logicalDrives, ov.LogicalDriveV3{
+					Bootable:          logicalDrivesItem["bootable"].(bool),
+					RaidLevel:         logicalDrivesItem["raid_level"].(string),
+					Accelerator:       logicalDrivesItem["accelerator"].(string),
+					DriveTechnology:   logicalDrivesItem["drive_technology"].(string),
+					Name:              logicalDrivesItem["name"].(string),
+					NumPhysicalDrives: logicalDrivesItem["num_physical_drives"].(int),
+					NumSpareDrives:    logicalDrivesItem["num_spare_drives"].(int),
+					SasLogicalJBODId:  logicalDrivesItem["sas_logical_jbod_id"].(int),
+				})
+			}
+			localStorageEmbeddedController = append(localStorageEmbeddedController, ov.LocalStorageEmbeddedController{
+				DeviceSlot:      controller_data["device_slot"].(string),
+				DriveWriteCache: controller_data["drive_write_cache"].(string),
+				Initialize:      controller_data["initialize"].(bool),
+				Mode:            controller_data["mode"].(string),
+				PredictiveSpareRebuild: controller_data["predictive_spare_rebuild"].(string),
+				LogicalDrives:          logicalDrives,
+			})
+		}
+		// Gets Local Storage Sas Jbods Body
+		rawLocalStorageSasJbod := localStorageItem["sas_logical_jbod"].(*schema.Set).List()
+		logicalJbod := make([]ov.LogicalJbod, 0)
+		for _, raw3 := range rawLocalStorageSasJbod {
+			sas_logical_jobd_data := raw3.(map[string]interface{})
+			logicalJbod = append(logicalJbod, ov.LogicalJbod{
+				Description:       sas_logical_jobd_data["description"].(string),
+				DeviceSlot:        sas_logical_jobd_data["drive_slot"].(string),
+				DriveMaxSizeGB:    sas_logical_jobd_data["drive_max_size_gb"].(int),
+				DriveMinSizeGB:    sas_logical_jobd_data["drive_min_size_gb"].(int),
+				DriveTechnology:   sas_logical_jobd_data["drive_technology"].(string),
+				EraseData:         sas_logical_jobd_data["erase_data"].(bool),
+				ID:                sas_logical_jobd_data["id"].(int),
+				Name:              sas_logical_jobd_data["name"].(string),
+				NumPhysicalDrives: sas_logical_jobd_data["num_physical_drive"].(int),
+				Persistent:        sas_logical_jobd_data["persistent"].(bool),
+			})
+		}
 		localStorage = ov.LocalStorageOptions{
 			ManageLocalStorage: localStorageItem["manage_local_storage"].(bool),
 			Initialize:         localStorageItem["initialize"].(bool),
+			Controllers:        localStorageEmbeddedController,
+			SasLogicalJBODs:    logicalJbod,
 		}
 	}
 	serverProfileTemplate.LocalStorage = localStorage
-
-	rawLogicalDrives := d.Get("logical_drives").(*schema.Set).List()
-	logicalDrives := make([]ov.LogicalDrive, 0)
-	for _, rawLogicalDrive := range rawLogicalDrives {
-		logicalDrivesItem := rawLogicalDrive.(map[string]interface{})
-		logicalDrives = append(logicalDrives, ov.LogicalDrive{
-			Bootable:  logicalDrivesItem["bootable"].(bool),
-			RaidLevel: logicalDrivesItem["raid_level"].(string),
-		})
-	}
-	serverProfileTemplate.LocalStorage.LogicalDrives = logicalDrives
 
 	// get SAN storage data if provided
 	rawSanStorage := d.Get("san_storage").(*schema.Set).List()
@@ -1007,6 +1146,56 @@ func resourceServerProfileTemplateRead(d *schema.ResourceData, meta interface{})
 		}
 		d.Set("boot_order", bootOrder)
 	}
+	// Gets Local Storage Body
+	local_storage := make([]map[string]interface{}, 0, 1)
+	// Gets Storage Controller Body
+	controllers := make([]map[string]interface{}, 0, len(spt.LocalStorage.Controllers))
+	for i := 0; i < len(spt.LocalStorage.Controllers); i++ {
+		logical_drives := make([]map[string]interface{}, 0, len(spt.LocalStorage.Controllers[i].LogicalDrives))
+		for j := 0; j < len(spt.LocalStorage.Controllers[i].LogicalDrives); j++ {
+			logical_drives = append(logical_drives, map[string]interface{}{
+				"bootable":            spt.LocalStorage.Controllers[i].LogicalDrives[j].Bootable,
+				"accelerator":         spt.LocalStorage.Controllers[i].LogicalDrives[j].Accelerator,
+				"drive_technology":    spt.LocalStorage.Controllers[i].LogicalDrives[j].DriveTechnology,
+				"name":                spt.LocalStorage.Controllers[i].LogicalDrives[j].Name,
+				"num_physical_drives": spt.LocalStorage.Controllers[i].LogicalDrives[j].NumPhysicalDrives,
+				"num_spare_drives":    spt.LocalStorage.Controllers[i].LogicalDrives[j].NumSpareDrives,
+				"sas_logical_jbod_id": spt.LocalStorage.Controllers[i].LogicalDrives[j].SasLogicalJBODId,
+				"raid_level":          spt.LocalStorage.Controllers[i].LogicalDrives[j].RaidLevel,
+			})
+		}
+		controllers = append(controllers, map[string]interface{}{
+			"device_slot":       spt.LocalStorage.Controllers[i].DeviceSlot,
+			"initialize":        spt.LocalStorage.Controllers[i].Initialize,
+			"drive_write_cache": spt.LocalStorage.Controllers[i].DriveWriteCache,
+			"mode":              spt.LocalStorage.Controllers[i].Mode,
+			"predictive_spare_rebuild": spt.LocalStorage.Controllers[i].PredictiveSpareRebuild,
+			"logical_drive":            logical_drives,
+		})
+	}
+	// Gets Sas Logical Jbod Controller Body
+	sas_logDrives := make([]map[string]interface{}, 0, len(spt.LocalStorage.SasLogicalJBODs))
+	for i := 0; i < len(spt.LocalStorage.SasLogicalJBODs); i++ {
+		sas_logDrives = append(sas_logDrives, map[string]interface{}{
+			"description":        spt.LocalStorage.SasLogicalJBODs[i].Description,
+			"device_slot":        spt.LocalStorage.SasLogicalJBODs[i].DeviceSlot,
+			"drive_max_size_gb":  spt.LocalStorage.SasLogicalJBODs[i].DriveMaxSizeGB,
+			"drive_min_size_sb":  spt.LocalStorage.SasLogicalJBODs[i].DriveMinSizeGB,
+			"drive_technology":   spt.LocalStorage.SasLogicalJBODs[i].DriveTechnology,
+			"erase_data":         spt.LocalStorage.SasLogicalJBODs[i].EraseData,
+			"id":                 spt.LocalStorage.SasLogicalJBODs[i].ID,
+			"name":               spt.LocalStorage.SasLogicalJBODs[i].Name,
+			"num_physical_drive": spt.LocalStorage.SasLogicalJBODs[i].NumPhysicalDrives,
+			"persistent":         spt.LocalStorage.SasLogicalJBODs[i].Persistent,
+		})
+	}
+	local_storage = append(local_storage, map[string]interface{}{
+		"manage_local_storage": spt.LocalStorage.ManageLocalStorage,
+		"initialize":           spt.LocalStorage.Initialize,
+		"controller":           controllers,
+		"sas_logical_jbod":     sas_logDrives,
+	})
+	d.Set("local_storage", local_storage)
 
 	return nil
 }
@@ -1125,29 +1314,84 @@ func resourceServerProfileTemplateUpdate(d *schema.ResourceData, meta interface{
 		}
 		serverProfileTemplate.Firmware = firmware
 	}
+
 	// Get local storage data if provided
 	rawLocalStorage := d.Get("local_storage").(*schema.Set).List()
 	localStorage := ov.LocalStorageOptions{}
 	for _, raw := range rawLocalStorage {
 		localStorageItem := raw.(map[string]interface{})
+		rawLocalStorageController := localStorageItem["controller"].(*schema.Set).List()
+		localStorageEmbeddedController := make([]ov.LocalStorageEmbeddedController, 0)
+		for _, raw2 := range rawLocalStorageController {
+			controller_data := raw2.(map[string]interface{})
+			rawLogicalDrives := controller_data["logical_drives"].(*schema.Set).List()
+			logicalDrives := make([]ov.LogicalDriveV3, 0)
+			for _, rawLogicalDrive := range rawLogicalDrives {
+				logicalDrivesItem := rawLogicalDrive.(map[string]interface{})
+				logicalDrives = append(logicalDrives, ov.LogicalDriveV3{
+					Bootable:          logicalDrivesItem["bootable"].(bool),
+					RaidLevel:         logicalDrivesItem["raid_level"].(string),
+					Accelerator:       logicalDrivesItem["accelerator"].(string),
+					DriveTechnology:   logicalDrivesItem["drive_technology"].(string),
+					Name:              logicalDrivesItem["name"].(string),
+					NumPhysicalDrives: logicalDrivesItem["num_physical_drives"].(int),
+					NumSpareDrives:    logicalDrivesItem["num_spare_drives"].(int),
+					SasLogicalJBODId:  logicalDrivesItem["sas_logical_jbod_id"].(int),
+				})
+			}
+			localStorageEmbeddedController = append(localStorageEmbeddedController, ov.LocalStorageEmbeddedController{
+				DeviceSlot:      controller_data["device_slot"].(string),
+				DriveWriteCache: controller_data["drive_write_cache"].(string),
+				Initialize:      controller_data["initialize"].(bool),
+				Mode:            controller_data["mode"].(string),
+				PredictiveSpareRebuild: controller_data["predictive_spare_rebuild"].(string),
+				LogicalDrives:          logicalDrives,
+			})
+		}
+		rawLocalStorageSasJbod := localStorageItem["sas_logical_jbod"].(*schema.Set).List()
+		logicalJbod := make([]ov.LogicalJbod, 0)
+		for _, raw3 := range rawLocalStorageSasJbod {
+			sas_logical_jobd_data := raw3.(map[string]interface{})
+			logicalJbod = append(logicalJbod, ov.LogicalJbod{
+				Description:       sas_logical_jobd_data["description"].(string),
+				DeviceSlot:        sas_logical_jobd_data["drive_slot"].(string),
+				DriveMaxSizeGB:    sas_logical_jobd_data["drive_max_size_gb"].(int),
+				DriveMinSizeGB:    sas_logical_jobd_data["drive_min_size_gb"].(int),
+				DriveTechnology:   sas_logical_jobd_data["drive_technology"].(string),
+				EraseData:         sas_logical_jobd_data["erase_data"].(bool),
+				ID:                sas_logical_jobd_data["id"].(int),
+				Name:              sas_logical_jobd_data["name"].(string),
+				NumPhysicalDrives: sas_logical_jobd_data["num_physical_drive"].(int),
+				Persistent:        sas_logical_jobd_data["persistent"].(bool),
+			})
+		}
 		localStorage = ov.LocalStorageOptions{
 			ManageLocalStorage: localStorageItem["manage_local_storage"].(bool),
 			Initialize:         localStorageItem["initialize"].(bool),
+			Controllers:        localStorageEmbeddedController,
+			SasLogicalJBODs:    logicalJbod,
 		}
 	}
 	serverProfileTemplate.LocalStorage = localStorage
 
-	rawLogicalDrives := d.Get("logical_drives").(*schema.Set).List()
-	logicalDrives := make([]ov.LogicalDrive, 0)
-	for _, rawLogicalDrive := range rawLogicalDrives {
-		logicalDrivesItem := rawLogicalDrive.(map[string]interface{})
-		logicalDrives = append(logicalDrives, ov.LogicalDrive{
-			Bootable:  logicalDrivesItem["bootable"].(bool),
-			RaidLevel: logicalDrivesItem["raid_level"].(string),
-		})
-	}
-	serverProfileTemplate.LocalStorage.LogicalDrives = logicalDrives
-
+	/*
+		rawLogicalDrives := d.Get("logical_drives").(*schema.Set).List()
+		logicalDrives := make([]ov.LogicalDriveV3, 0)
+		for _, rawLogicalDrive := range rawLogicalDrives {
+			logicalDrivesItem := rawLogicalDrive.(map[string]interface{})
+			logicalDrives = append(logicalDrives, ov.LogicalDriveV3{
+				Bootable:  logicalDrivesItem["bootable"].(bool),
+				RaidLevel: logicalDrivesItem["raid_level"].(string),
+				Accelerator: logicalDrivesItem["accelerator"].(string),
+				DriveTechnology: logicalDrivesItem["drive_technology"].(string),
+				Name:  logicalDrivesItem["name"].(string),
+				NumPhysicalDrives: logicalDrivesItem["num_physical_drives"].(int),
+				NumSpareDrives: logicalDrivesItem["num_spare_drives"].(int),
+				SasLogicalJBODId: logicalDrivesItem["sas_logical_jbod_id"].(int),
+			})
+		}
+		serverProfileTemplate.LocalStorage.LocalStorageEmbeddedController.LogicalDrives = logicalDrives
+	*/
 	// get SAN storage data if provided
 	rawSanStorage := d.Get("san_storage").(*schema.Set).List()
 	sanStorage := ov.SanStorageOptions{}