+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Bug fixes & Enhancements:
 - [#242] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/242) serverprofile bootmode_secureboot.
+- [#243] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/243) local storage configuration
 
 # [v1.6.0]
 ### Notes